requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
aiohttp==3.10.5
pydantic==2.9.2
pyyaml==6.0.2
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-multipart==0.0.12
